[ios] CB-2239: Support multiple plugin result values.
This allows a plugin to return multiple cordova exec bridge supported
plugin results.  The difference between a multipart message result and a
normal NSArray, is that a multipart message is "massaged" across the
bridge while an NSArray is just sent as-is.  This means that a multipart
message can include ArrayBuffers, while a normal NSArray plugin result
will not be scanned at all (possibly better perf).